Move routes enum to separate file
