+++ packages/client/src/routes.tsx
@@ -1,0 +1,11 @@
+enum routes {
+  mainPage = '/',
+  authPage = '/auth',
+  registerPage = '/register',
+  profilePage = '/profile',
+  liderboardPage = '/liderboard',
+  forumPage = '/forum',
+  gamePage = '/game'
+}
+
+export default routes;
